Revert "Update CRT submodules to latest releases (#1458)" (#1466)
This reverts commit f138efcaa33169b005cdbf5a0d11c10d89db292e.

As part of the investigation on the benchmark failures in the CI, e.g.
https://github.com/awslabs/mountpoint-s3/actions/runs/15625094824/job/44017689830,
we are reverting to the previous CRT releases.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Yes, reverted.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>