Introduce file rename support for directory buckets in S3 Express One Zone (#1468)
Introduces support in Mountpoint for renaming files, using the
[RenameObject
API](https://aws.amazon.com/about-aws/whats-new/2025/06/amazon-s3-express-one-zone-atomic-renaming-objects-api/),
which is supported on directory buckets in S3 Express One Zone.

File rename is enabled automatically when mounting a directory bucket in
S3 Express One Zone. In order to replace an existing object through
rename, the user must provide the `--allow-overwrite` flag at mount
time. More details on Mountpoint's support for rename can be found in
the semantics documentation `doc/SEMANTICS.md`.

### Does this change impact existing behavior?

Yes, this change will enable rename object when a bucket with support
for the new API is mounted.

### Does this change need a changelog entry? Does it require a version
change?

Changelog entries for the crates are updated. Versions are increased.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>