Update changelogs to prepare for crate release (#1470)
Updates the changelogs so that crates can be updated prior to the
release of MP v1.19.0

### Does this change impact existing behavior?

Does not impact behaviour.

### Does this change need a changelog entry? Does it require a version
change?

No, the crate update itself does not need a changelog entries. Version
changes were already done in #1468

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Christian Hagemeier <chagem@amazon.com>