Move object client into Superblock (#1476)
This PR moves the client into the Superblock, thus a superblock will
always interact with the same instantiation of an `ObjectClient + Send +
Sync`.

### Does this change impact existing behavior?

No, this change does not impact existing behaviour, as is only an
internal re-organisation.

### Does this change need a changelog entry? Does it require a version
change?

No, does not need a Changelog entry, as it only moves around where we
store the client.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>