Otlp exporter (#1464)
This PR adds an implementation of OpenTelemetry Exporting of metrics
through the OpenTelemetry protocol (OTLP). Changes are: a new
OtlpMetricsExporter struct which handles exporting metrics to an OTLP
endpoint, and integration of the OTLP exporter with the existing metrics
system.

Testing:
I tested the implementation with a test otlp_metrics() in metrics.rs and
ran a docker container running the OpenTelemetry Collector at the
default port

docker run -d --name otel-collector \
  -p 4318:4318 -p 4317:4317 \
  -v $(pwd)/collector-config.yaml:/etc/otelcol/config.yaml \
  otel/opentelemetry-collector-contrib:latest

Once I ran the test, I verified that the test metrics can be viewed in
the collector logs. (viewed using 'docker logs otel-collector'). Here is
a screenshot of an example of a test metric collected at the endpoint:
<img width="391" alt="Screenshot 2025-06-18 at 15 32 16"
src="https://github.com/user-attachments/assets/aab7e20a-0472-495b-af1d-23e966495e21"
/>

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Shivangi Kumar <shivyk@amazon.com>
Co-authored-by: Shivangi Kumar <shivyk@amazon.com>