Move `reader_count` out of inode (#1475)
Moves the reader count out of the inode and instead stores the reader
counts for all inodes with non-zero reader count in a HashMap (that is
protected by a lock).

### Does this change impact existing behavior?

This should not have breaking changes, it could potentially reduce
unlikely issues with the reader count getting messed up in highly
concurrent scenarios involving re-creation of inodes with the same
number.


### Does this change need a changelog entry? Does it require a version
change?

Does not need a Changelog entry or version change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>