Move lookup count into InodeMap (#1473)
This PR re-organises the way we lookup count by moving the lookup count
into inode HashMap (i.e, this hashmap now stores an association of
InodeID -> (inode, lookup_count)). This more closely mirrors real file
system's behaviour w.r.t. inodes that are re-created with the same inode
number. It introduces some additional locking.

This should not have any difference in behaviour, as we do not replace
inodes if they are currently open for writing or reading.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>