Add private option to disable disk cache cleanup for testing/benchmarking (#1483)
This updates the `ManagedCacheDir` struct to optionally perform cleanup
at creation and drop. It also provides an environment variable for
switching this on at `mount-s3` invocation time. This will allow us to
turn this cleanup off when trying to perform benchmarking comparing
"warming" phases, as well as "hot" phases - i.e. to understand how
Mountpoint performs when loading the cache versus a full cache, where
all requests are served from it.

### Does this change impact existing behavior?

No existing behavior change. One log is downgraded from `warn` to
`debug`. The warning was not providing much valuable information - if
needed, we can turn on debug logs if any strange behavior is observed.

### Does this change need a changelog entry? Does it require a version
change?

No, no public changes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>