Remove inodes from file handles  (#1486)
This PR removes the reference to the inodes from the file handle and
instead uses the inode number and full key.

### Does this change impact existing behavior?

No - is just an internal reorganisation.

### Does this change need a changelog entry? Does it require a version
change?

Needs no Changelog entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>
Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>
Co-authored-by: Alessandro Passaro <alexpax@amazon.co.uk>