Avoid copy of data returned by GetObject (#1481)
Reduce memory fragmentation and peak usage by avoiding copying data
returned by GetObject into newly allocated buffers. This change relies
on the new CRT API integrated in #1430, which allows `S3CrtClient` to
extend the lifetime of the buffers from the CRT memory pool when they
are returned by GetObject. Callers of the `get_object` method are now
responsible for dropping the returned `Bytes` instances in order for the
buffers to be released back to the CRT memory pool.

At the moment, the memory-limiting strategy used in the prefetcher
component in Mountpoint does not cope well with the change and may end
up starving the CRT of available buffers. For this specific use case, we
introduced a temporary feature flag in the `mountpoint-s3-client` crate,
`restore_buffer_copy` which restores the previous behavior, i.e.
GetObject allocates and returns new buffers with a copy of the object
content. As we rework this aspect of the prefetcher, we will likely
remove the feature flag.

### Does this change impact existing behavior?

Yes. The buffers returned by GetObject will be borrowed from the
internal memory pool.

### Does this change need a changelog entry? Does it require a version
change?

Yes. Entry and new version number for the client crate.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>