Add CLI option to publish metrics to an endpoint (#1478)
### Does this change impact existing behavior?
No impact on existing behavior, only adding on to existing behavior.

This PR adds a CLI argument which enables users to run Mountpoint with
the functionality of publishing metrics to a specified endpoint, and
includes a change to allow normal metrics logging and metrics publishing
at the same time.

Added functionality:
Run Mountpoint with `--log-metrics-otlp http://localhost:4318` flag to
enable publishing metrics to port 4318 (otlp port)
Optionally you can also specify the exporting interval with the
`--log-metrics-otlp-interval` flag.

To verify the implementation I ran a docker container running the
OpenTelemetry Collector at the default port, and ran Mountpoint with the
new flag with endpoint specified.
I verified that the Mountpoint metrics were visible in the collector
logs. Here is a screenshot of an example Mountpoint metric collected at
the endpoint:
<img width="408" alt="Screenshot 2025-06-24 at 17 23 38"
src="https://github.com/user-attachments/assets/a7acfdda-bb85-4595-96ea-633969418b79"
/>


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Shivangi Kumar <shivyk@amazon.com>
Co-authored-by: Shivangi Kumar <shivyk@amazon.com>