Remove clippy exception in logging module (#1497)
Minor internal change to the `logging` module in `mountpoint-s3-crt`:
when interfacing with the C functions, expose the logger implementation
as a direct reference instead of as a reference to a `Box`. Removes an
exception to the
[borrowed_box](https://rust-lang.github.io/rust-clippy/master/#borrowed_box)
clippy warning. Also adds the `unsafe` blocks and `SAFETY` comments that
will be required in Rust 2024.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>