Explicitly edit fstab file when running on Github runners to avoid a failure (#1482)
In the fstab CI tests, comment out a fstab entry for
`\dev/disk/cloud/azure_resource-part1` if we're running in Github
Actions.

### Does this change impact existing behavior?

Fixes a failure in Github CI.

### Does this change need a changelog entry? Does it require a version
change?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Simon Beal <simobeal@amazon.com>