Simplify lookup + adjust readdir interface (#1488)
This PR introduces a more generic Lookup structure and uses it as the
result type of a lookup.
Additionally adjusts the readdir interface.

### Does this change impact existing behavior?

No, does not impact existing behaviour.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>