Introduce builder pattern for mockclientconfig (#1502)
Use a builder pattern for MockClientConfig.

### Does this change impact existing behavior?

Does not impact existing behaviour as it only changes the way we build
the structure.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>