Introduce Metablock abstraction (#1500)
This PR introduces the `Metablock` abstraction that is currently only
implemented by the `Superblock`.
With this abstraction it will be easier to potentially introduce new
implementations of this interface for slightly modifed semantics.

Does not change existing behaviour, as it only introduces an interface.

Added Changelog entry for the `mountpoint-s3-fs` crate.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>