Extend prefetcher benchmarks to test multiple object downloads (#1504)
With this change, we can benchmark concurrent downloads of multiple
objects at prefetcher.

This change also allows passing NICs as a comma separated list and adds
a new parameter to limit the run time of the test

### Does this change impact existing behavior?

No, only extends prefetch benchmarks.

### Does this change need a changelog entry? Does it require a version
change?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Sahitya Damera <sahityad@amazon.com>