Update MP client_benchmark to support CRT backpressure (#1457)
Update MP client_backmark to support CRT backpressure. Extend the
benchmark to optionally enable read-backpressure in CRT, and configure
the initial read window size. This test aims to simulate the read-ahead
capability of the prefetcher, making it easier to baseline the
performance against the prefetcher benchmark.

### Does this change impact existing behavior?

No, changes to the benchmark only.

### Does this change need a changelog entry? Does it require a version
change?

No, changes to the benchmark only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Mansi Pandey <mansipnd@amazon.co.uk>
Signed-off-by: Christian Hagemeier <chagem@amazon.com>
Co-authored-by: Christian Hagemeier <chagem@amazon.com>