Add an option to write benchmarks output to a file (#1510)
This change allows us to save benchmarks output to a file in json
format, making it easier to parse the output.

### Does this change impact existing behavior?

No, prefetcher and client benchmarks only

### Does this change need a changelog entry? Does it require a version
change?

No, prefetcher and client benchmarks only

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Sahitya Damera <sahityad@amazon.com>