Refactor ClientBuilder into a trait and remove use of CliArgs (#1513)
We use a generic parameter in the `run` and `mount` functions to create
an S3 client instance (and associated runtime), so they can be used with
the actual S3 client and the mock one. This PR changes 2 things:
* Replaces the `FnOnce` with a trait, to make it simpler to pass around
and extend in the future,
* Removes the `CliArgs` argument in favor of `ClientConfig` and other
required settings.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>