Update MP client benchmarks to benchmark multiple object downloads  (#1512)
This change adds new parameters to client benchmarks to download
multiple objects and to limit the duration of the test.
This change also extends bind parameter to take a comma separated list
of NICs.

### Does this change impact existing behavior?

No, client benchmarks only

### Does this change need a changelog entry? Does it require a version
change?

No, client benchmarks only

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Sahitya Damera <sahityad@amazon.com>