Cherry-pick CRT updates (#1515, #1520) and prepare for client release 0.17.1 (#1521)
This change includes the fixes introduced in #1515 and #1520 and
increases the `mountpoint-s3-client` crate version to 0.17.1 in
preparation of publishing to [crates.io](https://crates.io).

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Yes. Cargo.toml and changelogs updated for patch release.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>