Add support for custom memory pools (#1516)
Introduces a `MemoryPool` trait in the client crate which allows users
to provide their own memory pool implementation. This is part of the
broader effort to use a unified memory pool in Mountpoint (see draft PR
#1511).

This change introduces:
* The required code to bridge implementations of the new Rust trait to
the CRT pool interface.
* A simple `MemoryPool` implementation to be used in tests.
* The `pool_tests` feature flags to use the above pool in the client
tests, replacing the CRT default pool.
* A new CI workflow to run the client tests with the custom pool.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Entry in the client changelog.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>