Extract meta request type when reserving buffers (#1524)
Modify the new `MemoryPool` trait (and the CRT bridge) to propagate the
type of the meta request which is reserving the buffer. Requires
extending bindings to an additional private header from `aws-c-s3` in
order to access the type of a `aws_s3_meta_request` pointer.

### Does this change impact existing behavior?

No, the new type information is not used yet.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>