Add custom memory pool implementation (#1529)
Introduce a custom implementation of a `MemoryPool` which can be used by
the CRT S3 client. The new pool will eventually be adopted in
Mountpoint, which will also use it to replace the allocations for disk
cache blocks and incremental upload buffers.

This change extends the integration tests on the client crate to run
with this pool implementation.

See docs in `memory/pool.rs` for more details on the new memory pool.

### Does this change impact existing behavior?

No, the new pool is only used in tests for now. 

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>