Remove duplicate S3Uri type (#1535)
The `S3Uri` was almost a complete duplicate of `S3Path`. This change
removes it and replaces it with `S3Path` in the few places where it was
used. It also rearranges the related validation methods and consolidates
types under the `s3` module.

**Note**: I split out the renames and moves into a separate commit for
ease of review.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Only for `fs` crate.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>