Fix issue preventing incremental upload to handle very large write part sizes (#1538)
The append upload queue tries to limit the total memory used to buffer
the data to write to 2 GiB. However, when setting `--write-part-size` to
values greater than 2 GiB, it would incorrectly set the queue capacity
to 0 buffers and panic.

This change ensures that the queue allows for at least 1 buffer, even if
that means exceeding the 2 GiB cap.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Bug fix entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>