Use a unified memory pool in Mountpoint (#1511)
Introduces a unified memory pool in Mountpoint. The pool adopts the CRT
pool interface, so it can be used by the CRT client when requesting
buffers. Ownership of the buffers is then passed to the prefetcher when
they are returned from GetObject requests. The same memory pool is also
used to serve reads from the local disk cache and for incremental
uploads.

The main goal is to reduce overall memory usage and mitigate memory
fragmentation issues. We may also observe performance gains in some
scenarios since we can avoid copying the data received from GetObject.

### Does this change impact existing behavior?

No changes in file system behavior. It will publish new memory-related
metrics.

### Does this change need a changelog entry? Does it require a version
change?

Requires entries in the `fs` and `mount-s3` changelogs and new major
versions.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>