Replace full key with S3Location in file handles (#1539)
Internal change to directly propagate `S3Location` in file handles
rather than the derived `full_key` string. The value is used for logging
and error report, so we can postpone formatting the string until it is
needed.

### Does this change impact existing behavior?

Minor change in string formatting in logs.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>