Enforce valid buffer sizes for the memory pool (#1540)
The memory pool will only accept buffer sizes in the range (0, 64MiB]
for the primary memory (i.e. allocated in pages of 16 buffers). For
larger sizes, it will only use secondary memory (i.e. ad-hoc allocation
for a single buffer).

The 64MiB cap reproduces the behavior of the internal CRT memory pool.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No - part of the memory pool change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>