Make ObjectClient part sizes no longer optional (#1542)
The `ObjectClient` trait currently defines `read_part_size` and
`write_part_size` as optional. This abstraction does not apply to any of
the existing implementations of the trait and we currently have no plans
of using it. This change removes this unnecessary abstraction,
simplifying the code and avoiding possible confusion.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Renan Magagnin <renanmag@amazon.co.uk>