Prepare for release of the fs crate v0.7.0 (#1544)
Update changelogs of the `fs` and `client` crates to prepare for
release.

Also include previously missing entry in `client` changelog for #1542,
and increase the crate version number.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

See above.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>