Use release flag for prefetcher benchmark (#1547)
Harmonises the use of `--release` compile time flag across benchmarks.

Does not need a Changelog entry, as it neither changes existing
behaviour nor is customer-facing.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Christian Hagemeier <chagem@amazon.com>