Explicitly specify opt-level, use link-time optimisations (#1548)
Explicitly set the optimisation level for our release builds to 3,
additionally enables link time optimisations and uses a single
compilation unit -- this enables more optimisations across the full
linked codebase.

Does not need a changelog entry, as it does not change mountpoint's
behaviour.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Christian Hagemeier <chagem@amazon.com>