Add versioning of the configuration format in mount_from_config example (#1545)
Example binary `mount_from_config` now accepts `config_version`
parameter. This may be used to ensure that user is aware of updates to
the configuration format and prevent from silent failures.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>