Update logging docs for CSI Driver v2 (#1551)
Update logging docs for CSI Driver v2 as in v2 logs are no longer in
syslog.

See
https://github.com/awslabs/mountpoint-s3-csi-driver/blob/main/docs/LOGGING.md

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Yerzhan Mazhkenov <20302932+yerzhan7@users.noreply.github.com>