Add helpful script for generating summary table from benchmark runs (#1557)
Adds a script that parses the benchmark output and autoamtically creates
a table with only the parameters that changed between runs.

Does not need a changelog entry, as the script only parses hydra runs. 


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>