Add option to disable download checksums in performance tests (#1555)
Adds an option to our benchmarking code to disable verification of
downloaded objects integrity.

Does not change existing behaviour, as it is only enabled when
`EXPERIMENTAL_MOUNTPOINT_NO_DOWNLOAD_INTEGRITY_VALIDATION ` is set, and
thus does not need a changelog entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>