Clean up the cache sub-directory in a background thread
Mountpoint tries to clean up the cache sub-directory during startup and exit in a blocking manner.
This causes a cold start on Mountpoint if the cache sub-directory is not cleaned up in exit (e.g., if it's forcefully terminated)
and if the cache sub-directory contains lots of files.

This change updates the clean up logic to do a rename first in the current thread to ensure Mountpoint starts with a fresh cache folder,
and then spawns a detached background thread to remove the old cache directory.
The cache sub-directory is still cleaned up in a blocking manner in the current thread during exit.

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>