Small fix for stdev of singleton list (#1572)
Fixes a small bug where the autogrouping script to analyse benchmark
runs in cmd may fail if a group only has 1 benchmark run, as
´statistics.stdev´ fails when the list has less than 2 entries. In these
cases, we will now return `N/A` as stddev.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Christian Hagemeier <chagem@amazon.com>