Enable resource monitoring for all benchmark types (#1573)
Adds the possibility to run resource monitoring for all benchmark types.
This is achieved by introducing a `Command` abstraction that is returned
by the benchmark. Then when the command has just started executing we
start the monitoring with it, unless we already have a PID to monitor
(used for FIO). (Thanks, Q )

Additionally changes the way we run most cargo commands by seperating
the phase where replacing `cargo run` by instead doing `cargo build` and
getting the executable path -- otherwise the compilation was part of the
flamegraph.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>