Capability to set EventLoopGroup thread count as a configurable CRT client config at runtime (#1579)
Add capability to set EventLoopGroup thread count as a configurable CRT
client config at runtime so we can override the current default with
different values (for e.g., during performance benchmarking) by setting
an environment variable `UNSTABLE_CRT_EVENTLOOP_THREADS`.

Note that the capability does not at the moment extend to CRT benchmarks
run through benchmark.py, because doing that will involve further
changes to etend CRT code and/or CRT benchmark logic.

Also note that unstable environmental variables are for experimental use
and may be removed or modified anytime.

### Does this change impact existing behavior?

No, it only introduces an unstable environment variable based capability
to _optionally_ configure the thread count for CRT's event loop group.
It also introduces a change in the MP/client/prefetcher benchmarks to
set that value using benchmark.py. The current default behaviours for
Mountpoint and benchmarks stay the same.

### Does this change need a changelog entry? Does it require a version
change?
No, as it's not a behavioural change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Mansi Pandey <mansipnd@amazon.com>
Co-authored-by: Mansi Pandey <mansipnd@amazon.com>