Remove locking assertion from unlink (#1596)
During unlink we currently have an assumption related to locking and we
assert it. However, we have seen some cases where the assumption does
not hold. The assumption is that, when removing the child node from the
parent node, the VFS will hold a lock on the parent and child.

This change removes the assumption and its assertion. Instead, we
invalidate the cache in the case where concurrent operations within the
same Mountpoint process were made to to the file and its parent.

For testing, we created scenarios to trigger the existing assertions in
the current implementation of unlink:
1. To trigger the `expect()` statement that follows the removal of the
`inode`:
- We added a 45s sleep statement in unlink between deletion from S3 and
the deletion from the `superblock`
- Workload: create a new file, remove it and (in a separate terminal)
executed a `stat` on the file
- The `stat` only completes after the deletion completes. The deletion
thread panics holding a lock and poisons the other threads. Mountpoint
unmounts.
2. To trigger the `assert()` statement that handles `inode` number
mismatch:
- We added a 45s sleep statement in unlink between deletion from S3 and
the deletion from the `superblock`
     - We added a 45s sleep statement in forget (`fs.rs`)
- Workload: create a new file, remove it, created a file with the same
name using `aws cli` and (in a separate terminal) a `stat` on the file
- Mountpoint behaviour is the same as the first case except the message
is from the assert.

With the changes in this PR, Mountpoint does not unmount and the `stat`
results are as expected (non existent in the first case and the most
recent file in the second case).

### Does this change impact existing behavior?

No, this is a fix.

### Does this change need a changelog entry? Does it require a version
change?

Yes, entires were added to the `CHANGELOG.md` files and the version of
the `mountpoint-s3-fs` crate was bumped to `0.7.1`.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Renan Magagnin <renanmag@amazon.co.uk>