Release v1.20.0 (#1604)
Update `CHANGELOG.md` files of the `fs` and `client` crates to prepare
for release.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

See above.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Renan Magagnin <renanmag@amazon.co.uk>