Remove benchmark config parser
As we have all default values in config.yaml and overrides in new hydra
sweeeper configuration files,  we don't need this additional layer of
parsing.

We verified the defaults remain unchanged except for
network.maximum_throughput_gbps, which was defaulting to 100 with the
parser. Now, it picks the default resolved by Mountpoint.

Signed-off-by: Sahitya Damera <sahityad@amazon.com>