Remove benchmark config parser (#1609)
As we shifted to using benchmark specific sweeper files, config.yaml
will define the defaults for all benchmark parameters and should not
have any overrides. So we don't need an additional parser to ensure
benchmarks are defaulting to appropriate values.

After this change, all the default values except for
network.maximum_throughput_gbps. The config parser was defaulting to 100
but now picks the default resolved by Mountpoint.

### Does this change impact existing behavior?

No benchmarks only

### Does this change need a changelog entry? Does it require a version
change?

No benchmarks only

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Sahitya Damera <sahityad@amazon.com>